+++ internal/executor/job/job_context.go
@@ -1,4 +1,4 @@
-package job_context
+package job
 
 import (
 	"sync"
