+++ internal/executor/job/submit_test.go
@@ -1,4 +1,4 @@
-package service
+package job
 
 import (
 	"testing"
