+++ internal/executor/service/job_lease_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/G-Research/armada/internal/common"
 	"github.com/G-Research/armada/internal/executor/configuration"
 	context2 "github.com/G-Research/armada/internal/executor/fake/context"
-	"github.com/G-Research/armada/internal/executor/job_context"
+	"github.com/G-Research/armada/internal/executor/job"
 	reporter_fake "github.com/G-Research/armada/internal/executor/reporter/fake"
 	"github.com/G-Research/armada/pkg/api"
 )
@@ -33,7 +33,7 @@ func TestCanBeRemovedConditions(t *testing.T) {
 	}
 
 	for pod, expected := range pods {
-		result := s.canBeRemoved(&job_context.RunningJob{JobId: "", Pods: []*v1.Pod{pod}})
+		result := s.canBeRemoved(&job.RunningJob{JobId: "", Pods: []*v1.Pod{pod}})
 		assert.Equal(t, expected, result)
 	}
 }
@@ -53,15 +53,15 @@ func TestCanBeRemovedMinumumPodTime(t *testing.T) {
 	}
 
 	for pod, expected := range pods {
-		result := s.canBeRemoved(&job_context.RunningJob{JobId: "", Pods: []*v1.Pod{pod}})
+		result := s.canBeRemoved(&job.RunningJob{JobId: "", Pods: []*v1.Pod{pod}})
 		assert.Equal(t, expected, result)
 	}
 }
 
 func TestChunkPods(t *testing.T) {
-	j := &job_context.RunningJob{}
-	chunks := chunkJobs([]*job_context.RunningJob{j, j, j}, 2)
-	assert.Equal(t, [][]*job_context.RunningJob{{j, j}, {j}}, chunks)
+	j := &job.RunningJob{}
+	chunks := chunkJobs([]*job.RunningJob{j, j, j}, 2)
+	assert.Equal(t, [][]*job.RunningJob{{j, j}, {j}}, chunks)
 }
 
 func makeFinishedPodWithTimestamp(state v1.PodPhase, timestamp time.Time) *v1.Pod {
@@ -91,7 +91,7 @@ func makePodWithCurrentStateReported(state v1.PodPhase, reportedDone bool) *v1.P
 
 func createLeaseService(minimumPodAge, failedPodExpiry time.Duration) *JobLeaseService {
 	fakeClusterContext := context2.NewFakeClusterContext(configuration.ApplicationConfiguration{ClusterId: "test", Pool: "pool"}, nil)
-	jobContext := job_context.NewClusterJobContext(fakeClusterContext)
+	jobContext := job.NewClusterJobContext(fakeClusterContext)
 	return NewJobLeaseService(fakeClusterContext, jobContext, &reporter_fake.FakeEventReporter{}, &queueClientMock{}, minimumPodAge, failedPodExpiry, common.ComputeResources{})
 }
